+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 * New: Disable WordPress Automatic Updates by default.
 * New: Option to limit the number of revisions.
+* New: Disable comments completely as default.
 * Fixed: Option Override Default Query for Post Type / Taxonomy is not working.
 
 ## 1.2.1 - 2021-09-23